Exception handled for Product service

+++ src/main/java/com/amazingcode/in/example/service/impl/ProductServiceImpl.java
@@ -1,10 +1,13 @@
 package com.amazingcode.in.example.service.impl;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
 import com.amazingcode.in.example.entity.Product;
+import com.amazingcode.in.example.exception.AlreadyPresentException;
+import com.amazingcode.in.example.exception.NotPresentException;
 import com.amazingcode.in.example.repository.ProductRepository;
 import com.amazingcode.in.example.service.ProductService;
 
@@ -19,24 +22,36 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public Product createProduct(Product product) {
+		boolean isProductPresent = productRepository.existsByProductName(product.getProductName());
+		if(isProductPresent) {
+			throw new AlreadyPresentException("Product with name "+product.getProductName()+" is already present.");
+		}
 		return productRepository.save(product);
 	}
 
 	@Override
 	public List<Product> getAllProducts() {
-		return productRepository.findAll();
+		List<Product> existsProducts = productRepository.findAll();
+		if(existsProducts.isEmpty()||existsProducts==null) {
+			throw new NotPresentException("Products not present.");
+		}
+		return existsProducts;
 	}
 
 	@Override
 	public Product getProduct(Long id) {
-		return productRepository.findById(id).get();
+		Optional<Product> existProduct = productRepository.findById(id);
+		if(existProduct.isEmpty()) {
+			throw new NotPresentException("Product with id "+id+" does not present.");
+		}
+		return existProduct.get();
 	}
 
 	@Override
 	public Product updateProduct(Long id, Product product) {
-		Product existProduct = productRepository.findById(id).get();
-		if (existProduct == null) {
-			return null;
+		Optional<Product> existProduct = productRepository.findById(id);
+		if(existProduct.isEmpty()) {
+			throw new NotPresentException("Product with id "+id+" does not present.");
 		}
 		product.setProductId(id);
 		Product updatedProduct = productRepository.save(product);
@@ -45,6 +60,10 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public void deleteProduct(Long id) {
+		Optional<Product> existProduct = productRepository.findById(id);
+		if(existProduct.isEmpty()) {
+			throw new NotPresentException("Product with id "+id+" does not exist to delete.");
+		}
 		productRepository.deleteById(id);
 	}
